# internal/api/openapi.yaml
openapi: 3.0.3
info:
  title: FlowForge API
  version: 0.1.0
  description: Durable sequential workflow execution. Authentication uses an optional API_TOKEN Bearer token; configure it whenever exposing the API beyond loopback. UTF-8 JSON request bodies are limited to 1 MiB.
servers:
  - url: /
security:
  - bearerAuth: []
paths:
  /health:
    get:
      operationId: health
      security: []
      responses:
        '200':
          description: Process is alive.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Health' }
  /ready:
    get:
      operationId: readiness
      security: []
      responses:
        '200':
          description: Required API dependencies are available, including PostgreSQL and applied migrations. Persisting new executions does not require a live broker.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Health' }
        '503': { $ref: '#/components/responses/Unavailable' }
  /api/v1/workflows:
    get:
      operationId: listWorkflows
      responses:
        '200':
          description: Up to 1000 immutable workflow definitions, ordered by creation time. Pagination is planned.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Workflow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '500': { $ref: '#/components/responses/InternalError' }
    post:
      operationId: createWorkflow
      parameters:
        - { $ref: '#/components/parameters/RequestID' }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CreateWorkflow' }
      responses:
        '201':
          description: Definition created. Definitions are immutable; create a new version instead of updating one.
          headers:
            Location: { schema: { type: string } }
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workflow' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/workflows/{id}:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    get:
      operationId: getWorkflow
      responses:
        '400': { $ref: '#/components/responses/ValidationError' }
        '200':
          description: Workflow definition.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Workflow' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/workflows/{id}/executions:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    post:
      operationId: createExecution
      description: Reusing an Idempotency-Key with the same workflow and canonical JSON input returns the original execution; a different workflow or payload returns 409. Keys are global and retained with their executions.
      parameters:
        - { $ref: '#/components/parameters/RequestID' }
        - name: Idempotency-Key
          in: header
          required: false
          schema: { type: string, minLength: 1, maxLength: 128, pattern: '^[!-~]+$' }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [input]
              properties:
                input: { type: object, additionalProperties: true }
      responses:
        '201':
          description: Execution persisted and available for scheduling.
          headers:
            Location: { schema: { type: string } }
            X-Request-ID: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Execution' }
        '200':
          description: Original execution returned for an idempotent replay.
          headers:
            Idempotency-Replayed: { schema: { type: string, enum: ['true'] } }
            Location: { schema: { type: string } }
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Execution' }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/TooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/executions/{id}:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    get:
      operationId: getExecution
      responses:
        '400': { $ref: '#/components/responses/ValidationError' }
        '200':
          description: Current persisted execution and step state.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Execution' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
  /api/v1/executions/{id}/events:
    parameters:
      - { $ref: '#/components/parameters/ID' }
    get:
      operationId: listExecutionEvents
      responses:
        '400': { $ref: '#/components/responses/ValidationError' }
        '200':
          description: Up to 10000 append-only execution history events, ordered by event ID. Pagination is planned.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Event' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '404': { $ref: '#/components/responses/NotFound' }
        '500': { $ref: '#/components/responses/InternalError' }
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer }
  parameters:
    ID:
      name: id
      in: path
      required: true
      schema: { type: string, format: uuid }
    RequestID:
      name: X-Request-ID
      in: header
      required: false
      description: A bounded visible ASCII correlation ID. Missing or invalid IDs are replaced with a UUID.
      schema: { type: string, maxLength: 128, pattern: '^[!-~]+$' }
  responses:
    ValidationError:
      description: A field or JSON body is invalid.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unauthorized:
      description: A Bearer token is required when API_TOKEN is configured.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    NotFound:
      description: The requested workflow or execution does not exist.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Conflict:
      description: The name/version pair or idempotency key conflicts with an existing resource.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    TooLarge:
      description: Request body exceeds 1 MiB.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    UnsupportedMediaType:
      description: Content-Type must be application/json.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    InternalError:
      description: An internal failure occurred; internal causes are not exposed.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
    Unavailable:
      description: A dependency is temporarily unavailable.
      headers:
        Retry-After: { schema: { type: string, example: '1' } }
      content:
        application/json:
          schema: { $ref: '#/components/schemas/Error' }
  schemas:
    Health:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [ok, ready] }
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, request_id]
          properties:
            code: { type: string }
            message: { type: string }
            request_id: { type: string }
    CreateWorkflow:
      type: object
      additionalProperties: false
      required: [name, version, steps]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        version: { type: integer, minimum: 1, maximum: 1000000 }
        description: { type: string, maxLength: 2000 }
        steps:
          type: array
          minItems: 1
          maxItems: 100
          items: { $ref: '#/components/schemas/CreateStep' }
    CreateStep:
      type: object
      additionalProperties: false
      required: [name, task_type]
      properties:
        name: { type: string, minLength: 1, maxLength: 100 }
        task_type: { type: string, maxLength: 128, pattern: '^[a-z][a-z0-9_]*(\.[a-z][a-z0-9_]*)+$' }
        max_attempts: { type: integer, minimum: 1, maximum: 10, default: 3 }
        timeout_seconds: { type: integer, minimum: 1, maximum: 3600, default: 30 }
    Workflow:
      type: object
      required: [id, name, version, description, steps, created_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        version: { type: integer }
        description: { type: string }
        created_at: { type: string, format: date-time }
        steps:
          type: array
          items: { $ref: '#/components/schemas/StepDefinition' }
    StepDefinition:
      type: object
      required: [id, name, task_type, max_attempts, timeout_seconds]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        task_type: { type: string }
        max_attempts: { type: integer }
        timeout_seconds: { type: integer }
    Execution:
      type: object
      required: [id, workflow_id, workflow_version, status, input, correlation_id, revision, created_at, started_at, completed_at, steps]
      properties:
        id: { type: string, format: uuid }
        workflow_id: { type: string, format: uuid }
        workflow_version: { type: integer }
        status: { type: string, enum: [pending, running, waiting, completed, failed, cancelled] }
        input: { type: object, additionalProperties: true }
        correlation_id: { type: string }
        revision: { type: integer, format: int64 }
        created_at: { type: string, format: date-time }
        started_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        steps:
          type: array
          items: { $ref: '#/components/schemas/StepExecution' }
    StepExecution:
      type: object
      required: [id, name, task_type, position, status, attempt, max_attempts, timeout_seconds, input, output, error_code, started_at, completed_at, retry_at, deadline_at]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        task_type: { type: string }
        position: { type: integer, minimum: 0 }
        status: { type: string, enum: [pending, running, retry_wait, completed, failed, cancelled] }
        attempt: { type: integer, minimum: 0 }
        max_attempts: { type: integer }
        timeout_seconds: { type: integer }
        input: { nullable: true }
        output: { nullable: true }
        error_code: { type: string }
        started_at: { type: string, format: date-time, nullable: true }
        completed_at: { type: string, format: date-time, nullable: true }
        retry_at: { type: string, format: date-time, nullable: true }
        deadline_at: { type: string, format: date-time, nullable: true }
    Event:
      type: object
      required: [id, execution_id, type, step_id, attempt, occurred_at, data]
      properties:
        id: { type: integer, format: int64 }
        execution_id: { type: string, format: uuid }
        type: { type: string }
        step_id: { type: string }
        attempt: { type: integer }
        occurred_at: { type: string, format: date-time }
        data: { nullable: true }